@@ -12,25 +12,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.2",
-    "@angular/cdk": "~13.2.2",
-    "@angular/common": "~13.2.2",
-    "@angular/compiler": "~13.2.2",
-    "@angular/core": "~13.2.2",
-    "@angular/forms": "~13.2.2",
-    "@angular/material": "~13.2.2",
-    "@angular/platform-browser": "~13.2.2",
-    "@angular/platform-browser-dynamic": "~13.2.2",
-    "@angular/router": "~13.2.2",
+    "@angular/animations": "^14.1.2",
+    "@angular/cdk": "^14.1.2",
+    "@angular/common": "^14.1.2",
+    "@angular/compiler": "^14.1.2",
+    "@angular/core": "^14.1.2",
+    "@angular/forms": "^14.1.2",
+    "@angular/material": "^14.1.2",
+    "@angular/platform-browser": "^14.1.2",
+    "@angular/platform-browser-dynamic": "^14.1.2",
+    "@angular/router": "^14.1.2",
     "crypto-js": "~4.1.1",
     "rxjs": "~6.6.0",
     "tslib": "~2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.3",
-    "@angular/cli": "~13.2.3",
-    "@angular/compiler-cli": "~13.2.2",
+    "@angular-devkit/build-angular": "^14.1.2",
+    "@angular/cli": "^14.1.2",
+    "@angular/compiler-cli": "^14.1.2",
     "@types/crypto-js": "~4.0.2",
     "@types/jasmine": "~3.8.0",
     "@types/node": "~12.11.1",
@@ -41,6 +41,6 @@
     "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "~1.7.0",
     "sonar-scanner": "~3.1.0",
-    "typescript": "~4.5.4"
+    "typescript": "~4.7.4"
   }
 }
